Refactored SimpleSMT into PureSMT; wrote the backend relying on typed-process instead of the lower level process library.
