+++ src/Pirouette/SMT.hs
@@ -13,7 +13,7 @@
 -- | This is Pirouette's SMT solver interface; you probably won't need to import
 --  anything under the @SMT/@ folder explicitely unless you're trying to do some
 --  very specific things or declaring a language. In which case, you probably
---  want only "Pirouette.SMT.Base" and "Pirouette.SMT.SimpleSMT" to bring get the necessary
+--  want only "Pirouette.SMT.Base" and "Pirouette.SMT.PureSMT" to bring get the necessary
 --  classes and definitions in scope. Check "Language.Pirouette.PlutusIR.SMT" for an example.
 module Pirouette.SMT
   ( SolverT (..),
@@ -39,7 +39,7 @@ module Pirouette.SMT
     -- * Convenient re-exports
     Constraint (..),
     AtomicConstraint (..),
-    SimpleSMT.Result (..),
+    PureSMT.Result (..),
     module Base,
     Branch (..),
     LanguageSMTBranches (..),
@@ -58,7 +58,7 @@ import Pirouette.Monad
 import Pirouette.Monad.Logger
 import Pirouette.SMT.Base as Base
 import Pirouette.SMT.Constraints
-import qualified Pirouette.SMT.SimpleSMT as SimpleSMT
+import qualified PureSMT
 import Pirouette.SMT.Translation
 import Pirouette.Term.Syntax
 import qualified Pirouette.Term.Syntax.SystemF as R
@@ -87,9 +87,9 @@ class Solve lang (problem :: * -> * -> *) where
 -- | Solver monad for a specific solver, passed as a phantom type variable @s@ (refer to 'IsSolver' for more)
 --  to know the supported solvers. That's a phantom type variable used only to distinguish
 --  solver-specific operations, such as initialization
-newtype SolverT m a = SolverT {unSolverT :: ReaderT SimpleSMT.Solver m a}
+newtype SolverT m a = SolverT {unSolverT :: ReaderT PureSMT.Solver m a}
   deriving (Functor)
-  deriving newtype (Applicative, Monad, MonadReader SimpleSMT.Solver, MonadIO, MonadLogger, MonadFail, MonadWeightedList)
+  deriving newtype (Applicative, Monad, MonadReader PureSMT.Solver, MonadIO, MonadLogger, MonadFail, MonadWeightedList)
 
 instance MonadTrans SolverT where
   lift = SolverT . lift
@@ -104,14 +104,14 @@ deriving instance PirouetteReadDefs lang m => PirouetteReadDefs lang (SolverT m)
 
 -- | Runs a computation that requires a session with a solver. The first parameter is
 -- an action that launches a solver. Check 'cvc4_ALL_SUPPORTED' for an example.
-runSolverT :: forall m a. (MonadIO m) => IO SimpleSMT.Solver -> SolverT m a -> m a
+runSolverT :: forall m a. (MonadIO m) => IO PureSMT.Solver -> SolverT m a -> m a
 runSolverT s (SolverT comp) = liftIO s >>= runReaderT comp
 
 -- | Returns 'Sat', 'Unsat' or 'Unknown' for the current solver session.
-checkSat :: (MonadIO m) => SolverT m SimpleSMT.Result
+checkSat :: (MonadIO m) => SolverT m PureSMT.Result
 checkSat = do
   solver <- ask
-  liftIO $ SimpleSMT.check solver
+  liftIO $ PureSMT.check solver
 
 -- | Pushes the solver context, creating a checkpoint. This is useful if you plan to execute
 --  many queries that share definitions. A typical use pattern would be:
@@ -127,13 +127,13 @@ checkSat = do
 solverPush :: (MonadIO m) => SolverT m ()
 solverPush = do
   solver <- ask
-  liftIO $ SimpleSMT.push solver
+  liftIO $ PureSMT.push solver
 
 -- | Pops the current checkpoint, check 'solverPush' for an example.
 solverPop :: (MonadIO m) => SolverT m ()
 solverPop = do
   solver <- ask
-  liftIO $ SimpleSMT.pop solver
+  liftIO $ PureSMT.pop solver
 
 -- | Declare a single datatype in the current solver session.
 declareDatatype :: (LanguageSMT lang, MonadIO m) => Name -> TypeDef lang -> ExceptT String (SolverT m) [Name]
@@ -141,7 +141,7 @@ declareDatatype typeName (Datatype _ typeVariables _ cstrs) = do
   solver <- ask
   (constr', _) <- runWriterT $ mapM (constructorFromPIR typeVariables) cstrs
   liftIO $ do
-    SimpleSMT.declareDatatype
+    PureSMT.declareDatatype
       solver
       (toSmtName typeName)
       (map (toSmtName . fst) typeVariables)
@@ -163,7 +163,7 @@ declareUninterpretedFunction fnName FunDef {funTy} = do
   let (args, result) = R.tyFunArgs funTy
   (args', _) <- runWriterT $ mapM translateType args
   (result', _) <- runWriterT $ translateType result
-  _ <- liftIO $ SimpleSMT.declareFun solver (toSmtName fnName) args' result'
+  _ <- liftIO $ PureSMT.declareFun solver (toSmtName fnName) args' result'
   return fnName
 
 declareUninterpretedFunctions ::
@@ -197,7 +197,7 @@ declareVariable :: (LanguageSMT lang, MonadIO m) => Name -> Type lang -> ExceptT
 declareVariable varName varTy = do
   solver <- ask
   (tySExpr, _) <- runWriterT $ translateType varTy
-  liftIO $ void (SimpleSMT.declare solver (toSmtName varName) tySExpr)
+  liftIO $ void (PureSMT.declare solver (toSmtName varName) tySExpr)
 
 -- | Asserts a constraint; check 'Constraint' for more information
 -- | The functions 'assert' and 'assertNot' output a boolean,
@@ -205,38 +205,38 @@ declareVariable varName varTy = do
 -- or if a part was lost during the translation process.
 assert ::
   (MonadIO m) =>
-  SimpleSMT.SExpr ->
+  PureSMT.SExpr ->
   SolverT m ()
 assert expr =
   SolverT $
     ReaderT $ \solver -> do
-      liftIO $ SimpleSMT.assert solver expr
+      liftIO $ PureSMT.assert solver expr
 
 assertNot ::
   (MonadIO m) =>
-  SimpleSMT.SExpr ->
+  PureSMT.SExpr ->
   SolverT m ()
 assertNot expr =
   SolverT $
     ReaderT $ \solver -> do
       -- liftIO $ putStrLn $ "asserting not " ++ show expr
-      liftIO $ SimpleSMT.assert solver (SimpleSMT.not expr)
+      liftIO $ PureSMT.assert solver (PureSMT.not expr)
 
 getUnsatCore ::
   (MonadIO m) =>
   SolverT m [String]
 getUnsatCore = SolverT $
   ReaderT $ \solver ->
-    liftIO $ SimpleSMT.getUnsatCore solver
+    liftIO $ PureSMT.getUnsatCore solver
 
 getModel ::
   (MonadIO m) =>
   [Name] ->
-  SolverT m [(SimpleSMT.SExpr, SimpleSMT.Value)]
+  SolverT m [(PureSMT.SExpr, PureSMT.Value)]
 getModel names = SolverT $
   ReaderT $ \solver -> do
-    let exprs = map (SimpleSMT.symbol . toSmtName) names
-    liftIO $ SimpleSMT.getExprs solver exprs
+    let exprs = map (PureSMT.symbol . toSmtName) names
+    liftIO $ PureSMT.getExprs solver exprs
 
 -- Utility functions
