+++ src/Pirouette/SMT/Base.hs
@@ -4,7 +4,7 @@ module Pirouette.SMT.Base where
 
 import Control.Monad.IO.Class
 import Data.Void
-import qualified Pirouette.SMT.SimpleSMT as SimpleSMT
+import qualified PureSMT
 import Pirouette.Term.Syntax
 
 -- | Captures the languages that can be translated to SMTLIB; namelly,
@@ -13,14 +13,14 @@ import Pirouette.Term.Syntax
 -- This class is defined with @-XAllowAmbiguousTypes@ and therefore
 -- should be used with @-XTypeApplications@ whenever necessary.
 class (LanguageBuiltins lang) => LanguageSMT lang where
-  translateBuiltinType :: BuiltinTypes lang -> SimpleSMT.SExpr
-  translateBuiltinTerm :: BuiltinTerms lang -> [SimpleSMT.SExpr] -> Maybe SimpleSMT.SExpr
-  translateConstant :: Constants lang -> SimpleSMT.SExpr
+  translateBuiltinType :: BuiltinTypes lang -> PureSMT.SExpr
+  translateBuiltinTerm :: BuiltinTerms lang -> [PureSMT.SExpr] -> Maybe PureSMT.SExpr
+  translateConstant :: Constants lang -> PureSMT.SExpr
   isStuckBuiltin :: TermMeta lang meta -> Bool
 
 -- | Captures arbitrary types that can be translated to SMTLIB.
 class (Show t) => ToSMT t where
-  translate :: t -> SimpleSMT.SExpr
+  translate :: t -> PureSMT.SExpr
 
 instance ToSMT Void where
   translate = absurd
@@ -30,7 +30,7 @@ toSmtName :: Name -> String
 toSmtName = ("pir_" <>) . show . pretty
 
 instance ToSMT Name where
-  translate = SimpleSMT.symbol . toSmtName
+  translate = PureSMT.symbol . toSmtName
 
 type WithDebugMessages = Bool
 
@@ -38,10 +38,8 @@ type WithDebugMessages = Bool
 -- to handle datatypes. The boolean parameter controls debug messages.
 -- The "fmf-fun" options is much better at finding sat by constructing finite model of recursive functions,
 -- whereas "full-saturate-quant" makes a much better use of the universally quantified hints to find unsat.
-cvc4_ALL_SUPPORTED :: (MonadIO m) => WithDebugMessages -> m SimpleSMT.Solver
+cvc4_ALL_SUPPORTED :: (MonadIO m) => WithDebugMessages -> m PureSMT.Solver
 cvc4_ALL_SUPPORTED dbg = do
-  -- This generates a "Solver" which logs every interaction it has.
-  ml <- if dbg then Just <$> liftIO (SimpleSMT.newLogger 0) else return Nothing
-  s <- liftIO $ SimpleSMT.newSolver "cvc4" ["--lang=smt2", "--incremental", "--fmf-fun"] ml
-  liftIO $ SimpleSMT.setLogic s "ALL"
+  s <- liftIO $ PureSMT.launchSolverWithFinalizer "cvc4 --lang=smt2 --incremental --fmf-fun" dbg
+  liftIO $ PureSMT.setLogic s "ALL"
   return s
