+++ src/Pirouette/SMT/Translation.hs
@@ -11,7 +11,7 @@
 {-# HLINT ignore "Replace case with maybe" #-}
 
 -- | Translation of Pirouette syntactical categories into
--- smtlib through our copy of the SimpleSMT library.
+-- smtlib through our copy of the PureSMT library.
 module Pirouette.SMT.Translation where
 
 import Control.Monad.Except
@@ -19,7 +19,7 @@ import Control.Monad.Writer (Any (..), WriterT (..), tell)
 -- import Debug.Trace (trace)
 import Pirouette.Monad
 import Pirouette.SMT.Base
-import qualified Pirouette.SMT.SimpleSMT as SmtLib
+import qualified PureSMT
 import Pirouette.Term.Syntax
 import qualified Pirouette.Term.Syntax.SystemF as Raw
 
@@ -75,18 +75,18 @@ translateTypeBase ::
   forall lang m.
   (LanguageSMT lang, Monad m) =>
   TypeBase lang ->
-  m SmtLib.SExpr
+  m PureSMT.SExpr
 translateTypeBase (TyBuiltin pirType) = return $ translateBuiltinType @lang pirType
-translateTypeBase (TySig name) = return $ SmtLib.symbol (toSmtName name)
+translateTypeBase (TySig name) = return $ PureSMT.symbol (toSmtName name)
 
 translateType ::
   (LanguageSMT lang, ToSMT meta, Monad m) =>
   TypeMeta lang meta ->
-  TranslatorT m SmtLib.SExpr
+  TranslatorT m PureSMT.SExpr
 translateType (Raw.TyApp (Raw.Free ty) args) =
-  SmtLib.app <$> translateTypeBase ty <*> mapM translateType args
+  PureSMT.app <$> translateTypeBase ty <*> mapM translateType args
 translateType (Raw.TyApp (Raw.Bound (Raw.Ann ann) _index) args) =
-  SmtLib.app (SmtLib.symbol (toSmtName ann)) <$> mapM translateType args
+  PureSMT.app (PureSMT.symbol (toSmtName ann)) <$> mapM translateType args
 translateType x =
   throwError $ "Translate type to smtlib: cannot handle " <> show x
 
@@ -99,7 +99,7 @@ translateTerm ::
   (LanguageSMT lang, ToSMT meta, PirouetteReadDefs lang m) =>
   [Name] ->
   TermMeta lang meta ->
-  TranslatorT m SmtLib.SExpr
+  TranslatorT m PureSMT.SExpr
 translateTerm _ (Raw.Lam _ann _ty _term) =
   throwError "Translate term to smtlib: Lambda abstraction in term"
 translateTerm _ (Raw.Abs _ann _kind _term) =
@@ -111,7 +111,7 @@ translateTerm knownNames (Raw.App var args) = case var of
   Raw.Free Bottom ->
     throwError "translateApp: Bottom; unclear how to translate that. WIP"
   -- meta go to ToSMT
-  Raw.Meta h -> SmtLib.app (translate h) <$> mapM (translateArg knownNames) args
+  Raw.Meta h -> PureSMT.app (translate h) <$> mapM (translateArg knownNames) args
   -- constants and builtins go to LanguageSMT
   Raw.Free (Constant c) ->
     if null args
@@ -146,16 +146,16 @@ translateTerm knownNames (Raw.App var args) = case var of
           -- there must be exactly as many arguments as required
           guard (length argTys == length restArgs)
           -- finally build the term
-          SmtLib.app
-            <$> (SmtLib.as (SmtLib.symbol (toSmtName name)) <$> translateType resultTy)
+          PureSMT.app
+            <$> (PureSMT.as (PureSMT.symbol (toSmtName name)) <$> translateType resultTy)
             <*> mapM (translateArg knownNames) restArgs
-      -- SmtLib.app (SmtLib.symbol (toSmtName name)) <$> mapM (translateArg knownNames) restArgs
+      -- PureSMT.app (PureSMT.symbol (toSmtName name)) <$> mapM (translateArg knownNames) restArgs
       DFunDef _
         | name `notElem` knownNames ->
           throwError $ "translateApp: Unknown function '" <> show name <> "'"
         | otherwise -> do
           tell UsedSomeUFs
-          SmtLib.app (SmtLib.symbol (toSmtName name)) <$> mapM (translateArg knownNames) args
+          PureSMT.app (PureSMT.symbol (toSmtName name)) <$> mapM (translateArg knownNames) args
       DTypeDef _ ->
         throwError "translateApp: Type name in function name"
       -- DO NEVER TRY TO TRANSLATE THESE!!
@@ -169,7 +169,7 @@ translateArg ::
   (LanguageSMT lang, ToSMT meta, PirouetteReadDefs lang m) =>
   [Name] ->
   ArgMeta lang meta ->
-  TranslatorT m SmtLib.SExpr
+  TranslatorT m PureSMT.SExpr
 translateArg knownNames (Raw.TermArg term) = translateTerm knownNames term
 -- TODO: This case is known to create invalid SMT terms,
 -- since in SMT solver, application of a term to a type is not allowed.
@@ -183,7 +183,7 @@ constructorFromPIR ::
   (LanguageSMT builtins, ToSMT meta, Monad m) =>
   [typeVariable] ->
   (Name, TypeMeta builtins meta) ->
-  TranslatorT m (String, [(String, SmtLib.SExpr)])
+  TranslatorT m (String, [(String, PureSMT.SExpr)])
 constructorFromPIR tyVars (name, constructorType) = do
   -- Fields of product types must be named: we append ids to the constructor name
   let fieldNames = map (((toSmtName name ++ "_") ++) . show) [1 :: Int ..]
