+++ tests/unit/Pirouette/Term/SymbolicEvalSpec.hs
@@ -15,7 +15,7 @@ import Pirouette.Transformations ( elimEvenOddMutRec )
 import Test.Tasty
 import Test.Tasty.HUnit
 import qualified Pirouette.SMT as SMT
-import qualified Pirouette.SMT.SimpleSMT as SMT
+import qualified PureSMT as SMT
 
 import Pirouette.Term.SymbolicEvalUtils
 
@@ -40,11 +40,11 @@ incorrectnessExec program term inC outC = fmap fst $ mockPrtT $ do
   orderedDecls <- elimEvenOddMutRec decls
   flip runReaderT orderedDecls $ do
     pathsIncorrectness (const False) UserDeclaredConstraints {
-      udcInputs = [], 
-      udcOutputCond = outC, 
-      udcInputCond = inC, 
-      udcAdditionalDefs = pure [], 
-      udcAxioms = [] } 
+      udcInputs = [],
+      udcOutputCond = outC,
+      udcInputCond = inC,
+      udcAdditionalDefs = pure [],
+      udcAxioms = [] }
       term
 
 add1 :: (Program Ex, Term Ex)
@@ -84,7 +84,7 @@ topConditions :: (Constraint Ex, TermMeta Ex SymVar -> Constraint Ex)
 topConditions = (mempty, const mempty)
 
 tests :: [TestTree]
-tests = [ 
+tests = [
   testCase "add 1" $
     symbolicExec' add1 `satisfies` isSingleton,
   testCase "add 1, bot" $
